fix/helpers: always trigger commands for a route navigated to
This fixes a bug introduced in 29f2f69bab6d6532217ccdbae178ccb116d342c2,
where we wouldn't trigger the commands if the route hadn't changed. The
problem is that when the login cycle finishes, most of the time the
route doesn't change (since we were waiting for the cycle to finish) and
we need to run commands (e.g. confirm email, get the word span in the
profile dashboard, etc.).